+++ WLUWEngine/WLUWEngine/src/WTexture.cpp
@@ -133,7 +133,7 @@ void WLUW::WTexture::setColor(Uint8 red, Uint8 green, Uint8 blue)
     SDL_SetTextureColorMod(texture, red, green, blue);
 }
 
-void WLUW::WTexture::render(SDL_Renderer* renderer, int x, int y, float scale, SDL_Rect* clip, double angle, SDL_Point* center, SDL_RendererFlip flip)
+void WLUW::WTexture::render(SDL_Renderer* renderer, int x, int y, double scale, SDL_Rect* clip, double angle, SDL_Point* center, SDL_RendererFlip flip)
 {
     //Set rendering space and render to screen
     SDL_Rect renderQuad = { x, y, size.x * scale, size.y * scale };
